Server Might Work
I set up a python server with each session getting a different thread and each client corresponds to one session. The host of each session is the first user in the list of users. The javascript I wrote connects to the open server with a hard coded IP address located in the settings.json file which will not be committed to hide my IP.